pyproject.toml
[tool.poetry]
name = "homemind"
version = "0.1.0"
description = ""
authors =[]
packages = [{include = "homemind"}]

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.39.0"
langchain-ollama = "^0.2.0"
langchain-core = "^0.3.10"
langchain = "^0.3.4"
langgraph = "^0.2.39"
ollama = "^0.3.3"
langchain-openai = "^0.2.8"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
streamlit-app = "scripts.run_streamlit:main"